Datalake UT (#1347)
