+++ CHANGELOG.md
@@ -1,7 +1,11 @@
-## 2.2.1 (Unreleased)
+## 2.3.0 (Unreleased)
 **Bug Fixes**
 
 **Features**
+- Migrated to the latest [azblob SDK](https://pkg.go.dev/github.com/Azure/azure-sdk-for-go/sdk/storage/azblob).
+- Migrated to the latest [azfile SDK](https://pkg.go.dev/github.com/Azure/azure-sdk-for-go/sdk/storage/azfile).
+- Migrated from deprecated ADAL to MSAL through the latest [azidentity SDK](https://pkg.go.dev/github.com/Azure/azure-sdk-for-go/sdk/azidentity).
+- Added support for uploading blobs in cold and premium tier.
 
 ## 2.2.0 (2024-01-24)
 **Bug Fixes**
