+++ component/azstorage/azauthkey.go
@@ -34,32 +34,69 @@
 package azstorage
 
 import (
-	"github.com/Azure/azure-storage-fuse/v2/common/log"
+	"errors"
 
-	"github.com/Azure/azure-storage-azcopy/v10/azbfs"
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob"
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/service"
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azdatalake"
+	serviceBfs "github.com/Azure/azure-sdk-for-go/sdk/storage/azdatalake/service"
+	"github.com/Azure/azure-storage-fuse/v2/common/log"
 )
 
 // Verify that the Auth implement the correct AzAuth interfaces
-var _ azAuth = &azAuthBfsKey{}
+var _ azAuth = &azAuthBlobKey{}
+var _ azAuth = &azAuthDatalakeKey{}
 
 type azAuthKey struct {
 	azAuthBase
 }
 
-type azAuthBfsKey struct {
+type azAuthBlobKey struct {
 	azAuthKey
 }
 
-// GetCredential : Gets shared key based storage credentials for datalake
-func (azkey *azAuthBfsKey) getCredential() interface{} {
+// getServiceClient : returns shared key based service client for blob
+func (azkey *azAuthBlobKey) getServiceClient(stConfig *AzStorageConfig) (interface{}, error) {
 	if azkey.config.AccountKey == "" {
-		log.Err("azAuthBfsKey::getCredential : Shared key for account is empty, cannot authenticate user")
-		return nil
+		log.Err("azAuthBlobKey::getServiceClient : Shared key for account is empty, cannot authenticate user")
+		return nil, errors.New("shared key for account is empty, cannot authenticate user")
 	}
 
-	credential := azbfs.NewSharedKeyCredential(
-		azkey.config.AccountName,
-		azkey.config.AccountKey)
+	cred, err := azblob.NewSharedKeyCredential(azkey.config.AccountName, azkey.config.AccountKey)
+	if err != nil {
+		log.Err("azAuthBlobKey::getServiceClient : Failed to create shared key credential [%s]", err.Error())
+		return nil, err
+	}
 
-	return credential
+	svcClient, err := service.NewClientWithSharedKeyCredential(azkey.config.Endpoint, cred, getAzBlobServiceClientOptions(stConfig))
+	if err != nil {
+		log.Err("azAuthBlobKey::getServiceClient : Failed to create service client [%s]", err.Error())
+	}
+
+	return svcClient, err
+}
+
+type azAuthDatalakeKey struct {
+	azAuthKey
+}
+
+// getServiceClient : returns shared key based service client for datalake
+func (azkey *azAuthDatalakeKey) getServiceClient(stConfig *AzStorageConfig) (interface{}, error) {
+	if azkey.config.AccountKey == "" {
+		log.Err("azAuthDatalakeKey::getServiceClient : Shared key for account is empty, cannot authenticate user")
+		return nil, errors.New("shared key for account is empty, cannot authenticate user")
+	}
+
+	cred, err := azdatalake.NewSharedKeyCredential(azkey.config.AccountName, azkey.config.AccountKey)
+	if err != nil {
+		log.Err("azAuthDatalakeKey::getServiceClient : Failed to create shared key credential [%s]", err.Error())
+		return nil, err
+	}
+
+	svcClient, err := serviceBfs.NewClientWithSharedKeyCredential(azkey.config.Endpoint, cred, getAzDatalakeServiceClientOptions(stConfig))
+	if err != nil {
+		log.Err("azAuthDatalakeKey::getServiceClient : Failed to create service client [%s]", err.Error())
+	}
+
+	return svcClient, err
 }
