+++ component/azstorage/azauthsas.go
@@ -34,25 +34,20 @@
 package azstorage
 
 import (
-	"fmt"
+	"errors"
+	"strings"
 
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/service"
+	serviceBfs "github.com/Azure/azure-sdk-for-go/sdk/storage/azdatalake/service"
 	"github.com/Azure/azure-storage-fuse/v2/common/log"
-
-	"github.com/Azure/azure-storage-azcopy/v10/azbfs"
 )
 
 // Verify that the Auth implement the correct AzAuth interfaces
-var _ azAuth = &azAuthBfsSAS{}
+var _ azAuth = &azAuthBlobSAS{}
+var _ azAuth = &azAuthDatalakeSAS{}
 
 type azAuthSAS struct {
 	azAuthBase
-}
-
-// GetEndpoint : Gets the SAS endpoint
-func (azsas *azAuthSAS) getEndpoint() string {
-	return fmt.Sprintf("%s%s",
-		azsas.config.Endpoint,
-		azsas.config.SASKey)
 }
 
 // SetOption : Sets the sas key information for the SAS auth.
@@ -62,16 +57,45 @@ func (azsas *azAuthSAS) setOption(key, value string) {
 	}
 }
 
-type azAuthBfsSAS struct {
+// GetEndpoint : Gets the SAS endpoint
+func (azsas *azAuthSAS) getEndpoint() string {
+	return azsas.config.Endpoint + "?" + strings.TrimLeft(azsas.config.SASKey, "?")
+}
+
+type azAuthBlobSAS struct {
 	azAuthSAS
 }
 
-// GetCredential : Gets SAS based credentials for datralake
-func (azsas *azAuthBfsSAS) getCredential() interface{} {
+// getServiceClient : returns SAS based service client for blob
+func (azsas *azAuthBlobSAS) getServiceClient(stConfig *AzStorageConfig) (interface{}, error) {
 	if azsas.config.SASKey == "" {
-		log.Err("azAuthBfsSAS::getCredential : SAS key for account is empty, cannot authenticate user")
-		return nil
+		log.Err("azAuthBlobSAS::getServiceClient : SAS key for account is empty, cannot authenticate user")
+		return nil, errors.New("sas key for account is empty, cannot authenticate user")
 	}
 
-	return azbfs.NewAnonymousCredential()
+	svcClient, err := service.NewClientWithNoCredential(azsas.getEndpoint(), getAzBlobServiceClientOptions(stConfig))
+	if err != nil {
+		log.Err("azAuthBlobSAS::getServiceClient : Failed to create service client [%s]", err.Error())
+	}
+
+	return svcClient, err
+}
+
+type azAuthDatalakeSAS struct {
+	azAuthSAS
+}
+
+// getServiceClient : returns SAS based service client for datalake
+func (azsas *azAuthDatalakeSAS) getServiceClient(stConfig *AzStorageConfig) (interface{}, error) {
+	if azsas.config.SASKey == "" {
+		log.Err("azAuthDatalakeSAS::getServiceClient : SAS key for account is empty, cannot authenticate user")
+		return nil, errors.New("sas key for account is empty, cannot authenticate user")
+	}
+
+	svcClient, err := serviceBfs.NewClientWithNoCredential(azsas.getEndpoint(), getAzDatalakeServiceClientOptions(stConfig))
+	if err != nil {
+		log.Err("azAuthDatalakeSAS::getServiceClient : Failed to create service client [%s]", err.Error())
+	}
+
+	return svcClient, err
 }
