+++ component/azstorage/azauthspn.go
@@ -34,122 +34,91 @@
 package azstorage
 
 import (
-	"fmt"
-	"math/rand"
-	"os"
-	"time"
-
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"github.com/Azure/azure-sdk-for-go/sdk/azidentity"
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/service"
+	serviceBfs "github.com/Azure/azure-sdk-for-go/sdk/storage/azdatalake/service"
 	"github.com/Azure/azure-storage-fuse/v2/common/log"
-
-	"github.com/Azure/azure-storage-azcopy/v10/azbfs"
-	"github.com/Azure/go-autorest/autorest/adal"
-	"github.com/Azure/go-autorest/autorest/azure"
 )
 
 // Verify that the Auth implement the correct AzAuth interfaces
-var _ azAuth = &azAuthBfsSPN{}
+var _ azAuth = &azAuthBlobSPN{}
+var _ azAuth = &azAuthDatalakeSPN{}
 
 type azAuthSPN struct {
 	azAuthBase
+	azOAuthBase
 }
 
-func getNextExpiryTimerSPN(spt *adal.ServicePrincipalToken) time.Duration {
-	delay := time.Duration(5+rand.Intn(120)) * time.Second
-	return time.Until(spt.Token().Expires()) - delay
-}
+func (azspn *azAuthSPN) getTokenCredential() (azcore.TokenCredential, error) {
+	var cred azcore.TokenCredential
+	var err error
 
-func (azspn *azAuthSPN) getAADEndpoint() string {
-	if azspn.config.ActiveDirectoryEndpoint != "" {
-		return azspn.config.ActiveDirectoryEndpoint
-	}
-	return azure.PublicCloud.ActiveDirectoryEndpoint
-}
-
-// fetchToken : Generates a token based on the config
-func (azspn *azAuthSPN) fetchToken() (*adal.ServicePrincipalToken, error) {
-	//  Use the configured AAD endpoint for token generation
-	config, err := adal.NewOAuthConfig(azspn.getAADEndpoint(), azspn.config.TenantID)
-	if err != nil {
-		log.Err("AzAuthSPN::fetchToken : Failed to generate OAuth Config for SPN [%s]", err.Error())
-		return nil, err
-	}
-
-	//  Create the resource URL
-	resourceURL := azspn.config.AuthResource
-	if resourceURL == "" {
-		resourceURL = azspn.getEndpoint()
-	}
-
-	//  Generate the SPN token
-	var spt *adal.ServicePrincipalToken
+	clOpts := azspn.getAzIdentityClientOptions(&azspn.config)
 	if azspn.config.OAuthTokenFilePath != "" {
-		log.Trace("AzAuthSPN::fetchToken : Going for fedrated token flow.")
+		log.Trace("AzAuthSPN::getTokenCredential : Going for fedrated token flow")
 
-		tokenReader := func() (string, error) {
-			token, err := os.ReadFile(azspn.config.OAuthTokenFilePath)
-			if err != nil {
-				return "", fmt.Errorf("failed to read OAuth token file %s [%v]", azspn.config.OAuthTokenFilePath, err.Error())
-			}
-			return string(token), nil
-		}
-
-		spt, err = adal.NewServicePrincipalTokenFromFederatedTokenCallback(*config, azspn.config.ClientID, tokenReader, resourceURL)
+		cred, err = azidentity.NewWorkloadIdentityCredential(&azidentity.WorkloadIdentityCredentialOptions{
+			ClientOptions: clOpts,
+			ClientID:      azspn.config.ClientID,
+			TenantID:      azspn.config.TenantID,
+			TokenFilePath: azspn.config.OAuthTokenFilePath,
+		})
 		if err != nil {
-			log.Err("AzAuthSPN::fetchToken : Failed to generate token for SPN [%s]", err.Error())
+			log.Err("AzAuthSPN::getTokenCredential : Failed to generate token for SPN [%s]", err.Error())
 			return nil, err
 		}
 	} else {
-		spt, err = adal.NewServicePrincipalToken(*config, azspn.config.ClientID, azspn.config.ClientSecret, resourceURL)
+		log.Trace("AzAuthSPN::getTokenCredential : Using client secret for fetching token")
+
+		cred, err = azidentity.NewClientSecretCredential(azspn.config.TenantID, azspn.config.ClientID, azspn.config.ClientSecret, &azidentity.ClientSecretCredentialOptions{
+			ClientOptions: clOpts,
+		})
 		if err != nil {
-			log.Err("AzAuthSPN::fetchToken : Failed to generate token for SPN [%s]", err.Error())
+			log.Err("AzAuthSPN::getTokenCredential : Failed to generate token for SPN [%s]", err.Error())
 			return nil, err
 		}
 	}
 
-	return spt, nil
+	return cred, err
 }
 
-type azAuthBfsSPN struct {
+type azAuthBlobSPN struct {
 	azAuthSPN
 }
 
-// GetCredential : Get SPN based credentials for datalake
-func (azspn *azAuthBfsSPN) getCredential() interface{} {
-
-	spt, err := azspn.fetchToken()
+// getServiceClient : returns SPN based service client for blob
+func (azspn *azAuthBlobSPN) getServiceClient(stConfig *AzStorageConfig) (interface{}, error) {
+	cred, err := azspn.getTokenCredential()
 	if err != nil {
-		log.Err("azAuthBfsSPN::getCredential : Failed to fetch token for SPN [%s]", err.Error())
-		return nil
+		log.Err("azAuthBlobSPN::getServiceClient : Failed to get token credential from SPN [%s]", err.Error())
+		return nil, err
 	}
 
-	// Using token create the credential object, here also register a call back which refreshes the token
-	tc := azbfs.NewTokenCredential(spt.Token().AccessToken, func(tc azbfs.TokenCredential) time.Duration {
-		// spt, err = azspn.fetchToken()
-		// if err != nil {
-		// 	log.Err("azAuthBfsSPN::getCredential : Failed to fetch SPN token [%s]", err.Error())
-		// 	return 0
-		// }
-		for failCount := 0; failCount < 5; failCount++ {
-			err = spt.Refresh()
-			if err != nil {
-				log.Err("azAuthBfsSPN::getCredential : Failed to refresh token attempt %d [%s]", failCount, err.Error())
-				time.Sleep(time.Duration(rand.Intn(5)) * time.Second)
-				continue
-			}
+	svcClient, err := service.NewClient(azspn.config.Endpoint, cred, getAzBlobServiceClientOptions(stConfig))
+	if err != nil {
+		log.Err("azAuthBlobSPN::getServiceClient : Failed to create service client [%s]", err.Error())
+	}
 
-			// set the new token value
-			tc.SetToken(spt.Token().AccessToken)
-			log.Info("azAuthBfsSPN::getCredential : SPN Token retrieved")
-			log.Debug("azAuthBfsSPN::getCredential : Token: %s (%s)", spt.Token().AccessToken, spt.Token().Expires())
+	return svcClient, err
+}
 
-			// Get the next token slightly before the current one expires
-			return getNextExpiryTimerSPN(spt)
-			// Test code to expire token every 30 seconds
-			// return time.Until(time.Now()) + 30*time.Second
-		}
-		log.Err("azAuthBfsSPN::getCredential : Failed to refresh token bailing out.")
-		return 0
-	})
+type azAuthDatalakeSPN struct {
+	azAuthSPN
+}
 
-	return tc
+// getServiceClient : returns SPN based service client for datalake
+func (azspn *azAuthDatalakeSPN) getServiceClient(stConfig *AzStorageConfig) (interface{}, error) {
+	cred, err := azspn.getTokenCredential()
+	if err != nil {
+		log.Err("azAuthDatalakeSPN::getServiceClient : Failed to get token credential from SPN [%s]", err.Error())
+		return nil, err
+	}
+
+	svcClient, err := serviceBfs.NewClient(azspn.config.Endpoint, cred, getAzDatalakeServiceClientOptions(stConfig))
+	if err != nil {
+		log.Err("azAuthDatalakeSPN::getServiceClient : Failed to create service client [%s]", err.Error())
+	}
+
+	return svcClient, err
 }
