+++ component/azstorage/datalake_test.go
@@ -53,24 +53,28 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/blob"
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/blockblob"
+
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azdatalake/file"
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azdatalake/filesystem"
+	"github.com/Azure/azure-sdk-for-go/sdk/storage/azdatalake/service"
+
 	"github.com/Azure/azure-storage-fuse/v2/common"
 	"github.com/Azure/azure-storage-fuse/v2/common/log"
 	"github.com/Azure/azure-storage-fuse/v2/internal"
 	"github.com/Azure/azure-storage-fuse/v2/internal/handlemap"
 
-	"github.com/Azure/azure-storage-azcopy/v10/azbfs"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 )
 
 type datalakeTestSuite struct {
 	suite.Suite
-	assert       *assert.Assertions
-	az           *AzStorage
-	serviceUrl   azbfs.ServiceURL
-	containerUrl azbfs.FileSystemURL
-	config       string
-	container    string
+	assert          *assert.Assertions
+	az              *AzStorage
+	serviceClient   *service.Client
+	containerClient *filesystem.Client
+	config          string
+	container       string
 }
 
 func (s *datalakeTestSuite) SetupTest() {
@@ -122,17 +126,17 @@ func (s *datalakeTestSuite) setupTestHelper(configuration string, container stri
 	s.az.Start(ctx) // Note: Start->TestValidation will fail but it doesn't matter. We are creating the container a few lines below anyway.
 	// We could create the container before but that requires rewriting the code to new up a service client.
 
-	s.serviceUrl = s.az.storage.(*Datalake).Service // Grab the service client to do some validation
-	s.containerUrl = s.serviceUrl.NewFileSystemURL(s.container)
+	s.serviceClient = s.az.storage.(*Datalake).Service // Grab the service client to do some validation
+	s.containerClient = s.serviceClient.NewFileSystemClient(s.container)
 	if create {
-		s.containerUrl.Create(ctx)
+		s.containerClient.Create(ctx, nil)
 	}
 }
 
 func (s *datalakeTestSuite) tearDownTestHelper(delete bool) {
 	s.az.Stop()
 	if delete {
-		s.containerUrl.Delete(ctx)
+		s.containerClient.Delete(ctx, nil)
 	}
 }
 
@@ -201,9 +205,9 @@ func (s *datalakeTestSuite) TestListContainers() {
 	num := 10
 	prefix := generateContainerName()
 	for i := 0; i < num; i++ {
-		f := s.serviceUrl.NewFileSystemURL(prefix + fmt.Sprint(i))
-		f.Create(ctx)
-		defer f.Delete(ctx)
+		f := s.serviceClient.NewFileSystemClient(prefix + fmt.Sprint(i))
+		f.Create(ctx, nil)
+		defer f.Delete(ctx, nil)
 	}
 
 	containers, err := s.az.ListContainers()
@@ -233,8 +237,8 @@ func (s *datalakeTestSuite) TestCreateDir() {
 
 			s.assert.Nil(err)
 			// Directory should be in the account
-			dir := s.containerUrl.NewDirectoryURL(internal.TruncateDirName(path))
-			_, err = dir.GetProperties(ctx)
+			dir := s.containerClient.NewDirectoryClient(internal.TruncateDirName(path))
+			_, err = dir.GetProperties(ctx, nil)
 			s.assert.Nil(err)
 		})
 	}
@@ -253,8 +257,8 @@ func (s *datalakeTestSuite) TestDeleteDir() {
 
 			s.assert.Nil(err)
 			// Directory should not be in the account
-			dir := s.containerUrl.NewDirectoryURL(internal.TruncateDirName(path))
-			_, err = dir.GetProperties(ctx)
+			dir := s.containerClient.NewDirectoryClient(internal.TruncateDirName(path))
+			_, err = dir.GetProperties(ctx, nil)
 			s.assert.NotNil(err)
 		})
 	}
@@ -296,15 +300,15 @@ func (s *datalakeTestSuite) setupHierarchy(base string) (*list.List, *list.List,
 
 	// Validate the paths were setup correctly and all paths exist
 	for p := a.Front(); p != nil; p = p.Next() {
-		_, err := s.containerUrl.NewDirectoryURL(p.Value.(string)).GetProperties(ctx)
+		_, err := s.containerClient.NewDirectoryClient(p.Value.(string)).GetProperties(ctx, nil)
 		s.assert.Nil(err)
 	}
 	for p := ab.Front(); p != nil; p = p.Next() {
-		_, err := s.containerUrl.NewDirectoryURL(p.Value.(string)).GetProperties(ctx)
+		_, err := s.containerClient.NewDirectoryClient(p.Value.(string)).GetProperties(ctx, nil)
 		s.assert.Nil(err)
 	}
 	for p := ac.Front(); p != nil; p = p.Next() {
-		_, err := s.containerUrl.NewDirectoryURL(p.Value.(string)).GetProperties(ctx)
+		_, err := s.containerClient.NewDirectoryClient(p.Value.(string)).GetProperties(ctx, nil)
 		s.assert.Nil(err)
 	}
 	return a, ab, ac
@@ -322,12 +326,12 @@ func (s *datalakeTestSuite) TestDeleteDirHierarchy() {
 
 	// a paths should be deleted
 	for p := a.Front(); p != nil; p = p.Next() {
-		_, err = s.containerUrl.NewDirectoryURL(p.Value.(string)).GetProperties(ctx)
+		_, err = s.containerClient.NewDirectoryClient(p.Value.(string)).GetProperties(ctx, nil)
 		s.assert.NotNil(err)
 	}
 	ab.PushBackList(ac) // ab and ac paths should exist
 	for p := ab.Front(); p != nil; p = p.Next() {
-		_, err = s.containerUrl.NewDirectoryURL(p.Value.(string)).GetProperties(ctx)
+		_, err = s.containerClient.NewDirectoryClient(p.Value.(string)).GetProperties(ctx, nil)
 		s.assert.Nil(err)
 	}
 }
@@ -346,7 +350,7 @@ func (s *datalakeTestSuite) TestDeleteSubDirPrefixPath() {
 	// a paths under c1 should be deleted
 	for p := a.Front(); p != nil; p = p.Next() {
 		path := p.Value.(string)
-		_, err = s.containerUrl.NewDirectoryURL(path).GetProperties(ctx)
+		_, err = s.containerClient.NewDirectoryClient(path).GetProperties(ctx, nil)
 		if strings.HasPrefix(path, base+"/c1") {
 			s.assert.NotNil(err)
 		} else {
@@ -355,7 +359,7 @@ func (s *datalakeTestSuite) TestDeleteSubDirPrefixPath() {
 	}
 	ab.PushBackList(ac) // ab and ac paths should exist
 	for p := ab.Front(); p != nil; p = p.Next() {
-		_, err = s.containerUrl.NewDirectoryURL(p.Value.(string)).GetProperties(ctx)
+		_, err = s.containerClient.NewDirectoryClient(p.Value.(string)).GetProperties(ctx, nil)
 		s.assert.Nil(err)
 	}
 }
@@ -370,8 +374,8 @@ func (s *datalakeTestSuite) TestDeleteDirError() {
 	s.assert.NotNil(err)
 	s.assert.EqualValues(syscall.ENOENT, err)
 	// Directory should not be in the account
-	dir := s.containerUrl.NewDirectoryURL(name)
-	_, err = dir.GetProperties(ctx)
+	dir := s.containerClient.NewDirectoryClient(name)
+	_, err = dir.GetProperties(ctx, nil)
 	s.assert.NotNil(err)
 }
 
@@ -416,8 +420,8 @@ func (s *datalakeTestSuite) TestIsDirEmptyError() {
 	s.assert.False(empty) // Note: See comment in BlockBlob.List. BlockBlob behaves differently from Datalake
 
 	// Directory should not be in the account
-	dir := s.containerUrl.NewDirectoryURL(name)
-	_, err := dir.GetProperties(ctx)
+	dir := s.containerClient.NewDirectoryClient(name)
+	_, err := dir.GetProperties(ctx, nil)
 	s.assert.NotNil(err)
 }
 
@@ -551,8 +555,8 @@ func (s *datalakeTestSuite) TestReadDirError() {
 	s.assert.NotNil(err) // Note: See comment in BlockBlob.List. BlockBlob behaves differently from Datalake
 	s.assert.Empty(entries)
 	// Directory should not be in the account
-	dir := s.containerUrl.NewDirectoryURL(name)
-	_, err = dir.GetProperties(ctx)
+	dir := s.containerClient.NewDirectoryClient(name)
+	_, err = dir.GetProperties(ctx, nil)
 	s.assert.NotNil(err)
 }
 
@@ -597,13 +601,13 @@ func (s *datalakeTestSuite) TestRenameDir() {
 			err := s.az.RenameDir(internal.RenameDirOptions{Src: input.src, Dst: input.dst})
 			s.assert.Nil(err)
 			// Src should not be in the account
-			dir := s.containerUrl.NewDirectoryURL(internal.TruncateDirName(input.src))
-			_, err = dir.GetProperties(ctx)
+			dir := s.containerClient.NewDirectoryClient(internal.TruncateDirName(input.src))
+			_, err = dir.GetProperties(ctx, nil)
 			s.assert.NotNil(err)
 
 			// Dst should be in the account
-			dir = s.containerUrl.NewDirectoryURL(internal.TruncateDirName(input.dst))
-			_, err = dir.GetProperties(ctx)
+			dir = s.containerClient.NewDirectoryClient(internal.TruncateDirName(input.dst))
+			_, err = dir.GetProperties(ctx, nil)
 			s.assert.Nil(err)
 		})
 	}
@@ -624,23 +628,23 @@ func (s *datalakeTestSuite) TestRenameDirHierarchy() {
 	// Source
 	// aSrc paths should be deleted
 	for p := aSrc.Front(); p != nil; p = p.Next() {
-		_, err = s.containerUrl.NewDirectoryURL(p.Value.(string)).GetProperties(ctx)
+		_, err = s.containerClient.NewDirectoryClient(p.Value.(string)).GetProperties(ctx, nil)
 		s.assert.NotNil(err)
 	}
 	abSrc.PushBackList(acSrc) // abSrc and acSrc paths should exist
 	for p := abSrc.Front(); p != nil; p = p.Next() {
-		_, err = s.containerUrl.NewDirectoryURL(p.Value.(string)).GetProperties(ctx)
+		_, err = s.containerClient.NewDirectoryClient(p.Value.(string)).GetProperties(ctx, nil)
 		s.assert.Nil(err)
 	}
 	// Destination
 	// aDst paths should exist
 	for p := aDst.Front(); p != nil; p = p.Next() {
-		_, err = s.containerUrl.NewDirectoryURL(p.Value.(string)).GetProperties(ctx)
+		_, err = s.containerClient.NewDirectoryClient(p.Value.(string)).GetProperties(ctx, nil)
 		s.assert.Nil(err)
 	}
 	abDst.PushBackList(acDst) // abDst and acDst paths should not exist
 	for p := abDst.Front(); p != nil; p = p.Next() {
-		_, err = s.containerUrl.NewDirectoryURL(p.Value.(string)).GetProperties(ctx)
+		_, err = s.containerClient.NewDirectoryClient(p.Value.(string)).GetProperties(ctx, nil)
 		s.assert.NotNil(err)
 	}
 }
@@ -661,7 +665,7 @@ func (s *datalakeTestSuite) TestRenameDirSubDirPrefixPath() {
 	// aSrc paths under c1 should be deleted
 	for p := aSrc.Front(); p != nil; p = p.Next() {
 		path := p.Value.(string)
-		_, err = s.containerUrl.NewDirectoryURL(path).GetProperties(ctx)
+		_, err = s.containerClient.NewDirectoryClient(path).GetProperties(ctx, nil)
 		if strings.HasPrefix(path, baseSrc+"/c1") {
 			s.assert.NotNil(err)
 		} else {
@@ -670,14 +674,14 @@ func (s *datalakeTestSuite) TestRenameDirSubDirPrefixPath() {
 	}
 	abSrc.PushBackList(acSrc) // abSrc and acSrc paths should exist
 	for p := abSrc.Front(); p != nil; p = p.Next() {
-		_, err = s.containerUrl.NewDirectoryURL(p.Value.(string)).GetProperties(ctx)
+		_, err = s.containerClient.NewDirectoryClient(p.Value.(string)).GetProperties(ctx, nil)
 		s.assert.Nil(err)
 	}
 	// Destination
 	// aDst paths should exist -> aDst and aDst/gc1
-	_, err = s.containerUrl.NewDirectoryURL(baseSrc + "/" + baseDst).GetProperties(ctx)
+	_, err = s.containerClient.NewDirectoryClient(baseSrc+"/"+baseDst).GetProperties(ctx, nil)
 	s.assert.Nil(err)
-	_, err = s.containerUrl.NewDirectoryURL(baseSrc + "/" + baseDst + "/gc1").GetProperties(ctx)
+	_, err = s.containerClient.NewDirectoryClient(baseSrc+"/"+baseDst+"/gc1").GetProperties(ctx, nil)
 	s.assert.Nil(err)
 }
 
@@ -692,11 +696,11 @@ func (s *datalakeTestSuite) TestRenameDirError() {
 	s.assert.NotNil(err)
 	s.assert.EqualValues(syscall.ENOENT, err)
 	// Neither directory should be in the account
-	dir := s.containerUrl.NewDirectoryURL(src)
-	_, err = dir.GetProperties(ctx)
+	dir := s.containerClient.NewDirectoryClient(src)
+	_, err = dir.GetProperties(ctx, nil)
 	s.assert.NotNil(err)
-	dir = s.containerUrl.NewDirectoryURL(dst)
-	_, err = dir.GetProperties(ctx)
+	dir = s.containerClient.NewDirectoryClient(dst)
+	_, err = dir.GetProperties(ctx, nil)
 	s.assert.NotNil(err)
 }
 
@@ -712,11 +716,11 @@ func (s *datalakeTestSuite) TestCreateFile() {
 	s.assert.EqualValues(name, h.Path)
 	s.assert.EqualValues(0, h.Size)
 	// File should be in the account
-	file := s.containerUrl.NewDirectoryURL(name)
-	props, err := file.GetProperties(ctx)
+	file := s.containerClient.NewDirectoryClient(name)
+	props, err := file.GetProperties(ctx, nil)
 	s.assert.Nil(err)
 	s.assert.NotNil(props)
-	s.assert.Empty(props.XMsProperties())
+	s.assert.Empty(props.Metadata)
 }
 
 func (s *datalakeTestSuite) TestWriteSmallFile() {
@@ -1133,8 +1137,8 @@ func (s *datalakeTestSuite) TestDeleteFile() {
 	s.assert.Nil(err)
 
 	// File should not be in the account
-	file := s.containerUrl.NewDirectoryURL(name)
-	_, err = file.GetProperties(ctx)
+	file := s.containerClient.NewDirectoryClient(name)
+	_, err = file.GetProperties(ctx, nil)
 	s.assert.NotNil(err)
 }
 
@@ -1148,8 +1152,8 @@ func (s *datalakeTestSuite) TestDeleteFileError() {
 	s.assert.EqualValues(syscall.ENOENT, err)
 
 	// File should not be in the account
-	file := s.containerUrl.NewDirectoryURL(name)
-	_, err = file.GetProperties(ctx)
+	file := s.containerClient.NewDirectoryClient(name)
+	_, err = file.GetProperties(ctx, nil)
 	s.assert.NotNil(err)
 }
 
@@ -1164,12 +1168,12 @@ func (s *datalakeTestSuite) TestRenameFile() {
 	s.assert.Nil(err)
 
 	// Src should not be in the account
-	source := s.containerUrl.NewDirectoryURL(src)
-	_, err = source.GetProperties(ctx)
+	source := s.containerClient.NewDirectoryClient(src)
+	_, err = source.GetProperties(ctx, nil)
 	s.assert.NotNil(err)
 	// Dst should be in the account
-	destination := s.containerUrl.NewDirectoryURL(dst)
-	_, err = destination.GetProperties(ctx)
+	destination := s.containerClient.NewDirectoryClient(dst)
+	_, err = destination.GetProperties(ctx, nil)
 	s.assert.Nil(err)
 }
 
@@ -1177,27 +1181,26 @@ func (s *datalakeTestSuite) TestRenameFileMetadataConservation() {
 	defer s.cleanupTest()
 	// Setup
 	src := generateFileName()
-	source := s.containerUrl.NewRootDirectoryURL().NewFileURL(src)
+	source := s.containerClient.NewFileClient(src)
 	s.az.CreateFile(internal.CreateFileOptions{Name: src})
 	// Add srcMeta to source
-	srcMeta := make(map[string]string)
-	srcMeta["foo"] = "bar"
-	source.CreateWithOptions(ctx, azbfs.CreateFileOptions{Metadata: srcMeta}, azbfs.BlobFSAccessControl{})
+	srcMeta := make(map[string]*string)
+	srcMeta["foo"] = to.Ptr("bar")
+	source.SetMetadata(ctx, srcMeta, nil)
 	dst := generateFileName()
 
 	err := s.az.RenameFile(internal.RenameFileOptions{Src: src, Dst: dst})
 	s.assert.Nil(err)
 
 	// Src should not be in the account
-	_, err = source.GetProperties(ctx)
+	_, err = source.GetProperties(ctx, nil)
 	s.assert.NotNil(err)
 	// Dst should be in the account
-	destination := s.containerUrl.NewRootDirectoryURL().NewFileURL(dst)
-	props, err := destination.GetProperties(ctx)
+	destination := s.containerClient.NewFileClient(dst)
+	props, err := destination.GetProperties(ctx, nil)
 	s.assert.Nil(err)
 	// Dst should have metadata
-	destMeta := newMetadata(props.XMsProperties())
-	s.assert.True(checkMetadata(destMeta, "foo", "bar"))
+	s.assert.True(checkMetadata(props.Metadata, "foo", "bar")) //TODO:: track2: review
 }
 
 func (s *datalakeTestSuite) TestRenameFileError() {
@@ -1211,11 +1214,11 @@ func (s *datalakeTestSuite) TestRenameFileError() {
 	s.assert.EqualValues(syscall.ENOENT, err)
 
 	// Src and destination should not be in the account
-	source := s.containerUrl.NewDirectoryURL(src)
-	_, err = source.GetProperties(ctx)
+	source := s.containerClient.NewDirectoryClient(src)
+	_, err = source.GetProperties(ctx, nil)
 	s.assert.NotNil(err)
-	destination := s.containerUrl.NewDirectoryURL(dst)
-	_, err = destination.GetProperties(ctx)
+	destination := s.containerClient.NewDirectoryClient(dst)
+	_, err = destination.GetProperties(ctx, nil)
 	s.assert.NotNil(err)
 }
 
@@ -1328,10 +1331,12 @@ func (s *datalakeTestSuite) TestWriteFile() {
 	s.assert.EqualValues(len(data), count)
 
 	// Blob should have updated data
-	file := s.containerUrl.NewRootDirectoryURL().NewFileURL(name)
-	resp, err := file.Download(ctx, 0, int64(len(data)))
+	fileClient := s.containerClient.NewFileClient(name)
+	resp, err := fileClient.DownloadStream(ctx, &file.DownloadStreamOptions{
+		Range: &file.HTTPRange{Offset: 0, Count: int64(len(data))},
+	})
 	s.assert.Nil(err)
-	output, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body)
 	s.assert.EqualValues(testData, output)
 }
 
@@ -1349,11 +1354,14 @@ func (s *datalakeTestSuite) TestTruncateSmallFileSmaller() {
 	s.assert.Nil(err)
 
 	// Blob should have updated data
-	file := s.containerUrl.NewRootDirectoryURL().NewFileURL(name)
-	resp, err := file.Download(ctx, 0, int64(truncatedLength))
+	fileClient := s.containerClient.NewFileClient(name)
+	resp, err := fileClient.DownloadStream(ctx, &file.DownloadStreamOptions{
+		Range: &file.HTTPRange{Offset: 0, Count: int64(truncatedLength)},
+	})
+	// 0, int64(truncatedLength))
 	s.assert.Nil(err)
-	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	s.assert.EqualValues(truncatedLength, *resp.ContentLength)
+	output, _ := io.ReadAll(resp.Body)
 	s.assert.EqualValues(testData[:truncatedLength], output)
 }
 
@@ -1376,11 +1384,13 @@ func (s *datalakeTestSuite) TestTruncateChunkedFileSmaller() {
 	s.assert.Nil(err)
 
 	// Blob should have updated data
-	file := s.containerUrl.NewRootDirectoryURL().NewFileURL(name)
-	resp, err := file.Download(ctx, 0, int64(truncatedLength))
+	fileClient := s.containerClient.NewFileClient(name)
+	resp, err := fileClient.DownloadStream(ctx, &file.DownloadStreamOptions{
+		Range: &file.HTTPRange{Offset: 0, Count: int64(truncatedLength)},
+	})
 	s.assert.Nil(err)
-	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	s.assert.EqualValues(truncatedLength, *resp.ContentLength)
+	output, _ := io.ReadAll(resp.Body)
 	s.assert.EqualValues(testData[:truncatedLength], output)
 }
 
@@ -1398,11 +1408,13 @@ func (s *datalakeTestSuite) TestTruncateSmallFileEqual() {
 	s.assert.Nil(err)
 
 	// Blob should have updated data
-	file := s.containerUrl.NewRootDirectoryURL().NewFileURL(name)
-	resp, err := file.Download(ctx, 0, int64(truncatedLength))
+	fileClient := s.containerClient.NewFileClient(name)
+	resp, err := fileClient.DownloadStream(ctx, &file.DownloadStreamOptions{
+		Range: &file.HTTPRange{Offset: 0, Count: int64(truncatedLength)},
+	})
 	s.assert.Nil(err)
-	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	s.assert.EqualValues(truncatedLength, *resp.ContentLength)
+	output, _ := io.ReadAll(resp.Body)
 	s.assert.EqualValues(testData, output)
 }
 
@@ -1425,11 +1437,13 @@ func (s *datalakeTestSuite) TestTruncateChunkedFileEqual() {
 	s.assert.Nil(err)
 
 	// Blob should have updated data
-	file := s.containerUrl.NewRootDirectoryURL().NewFileURL(name)
-	resp, err := file.Download(ctx, 0, int64(truncatedLength))
+	fileClient := s.containerClient.NewFileClient(name)
+	resp, err := fileClient.DownloadStream(ctx, &file.DownloadStreamOptions{
+		Range: &file.HTTPRange{Offset: 0, Count: int64(truncatedLength)},
+	})
 	s.assert.Nil(err)
-	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	s.assert.EqualValues(truncatedLength, *resp.ContentLength)
+	output, _ := io.ReadAll(resp.Body)
 	s.assert.EqualValues(testData, output)
 }
 
@@ -1447,11 +1461,13 @@ func (s *datalakeTestSuite) TestTruncateSmallFileBigger() {
 	s.assert.Nil(err)
 
 	// Blob should have updated data
-	file := s.containerUrl.NewRootDirectoryURL().NewFileURL(name)
-	resp, err := file.Download(ctx, 0, int64(truncatedLength))
+	fileClient := s.containerClient.NewFileClient(name)
+	resp, err := fileClient.DownloadStream(ctx, &file.DownloadStreamOptions{
+		Range: &file.HTTPRange{Offset: 0, Count: int64(truncatedLength)},
+	})
 	s.assert.Nil(err)
-	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	s.assert.EqualValues(truncatedLength, *resp.ContentLength)
+	output, _ := io.ReadAll(resp.Body)
 	s.assert.EqualValues(testData, output[:len(data)])
 }
 
@@ -1474,11 +1490,13 @@ func (s *datalakeTestSuite) TestTruncateChunkedFileBigger() {
 	s.assert.Nil(err)
 
 	// Blob should have updated data
-	file := s.containerUrl.NewRootDirectoryURL().NewFileURL(name)
-	resp, err := file.Download(ctx, 0, int64(truncatedLength))
+	fileClient := s.containerClient.NewFileClient(name)
+	resp, err := fileClient.DownloadStream(ctx, &file.DownloadStreamOptions{
+		Range: &file.HTTPRange{Offset: 0, Count: int64(truncatedLength)},
+	})
 	s.assert.Nil(err)
-	s.assert.EqualValues(truncatedLength, resp.ContentLength())
-	output, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	s.assert.EqualValues(truncatedLength, *resp.ContentLength)
+	output, _ := io.ReadAll(resp.Body)
 	s.assert.EqualValues(testData, output[:len(data)])
 }
 
@@ -1544,10 +1562,12 @@ func (s *datalakeTestSuite) TestCopyFromFile() {
 	s.assert.Nil(err)
 
 	// Blob should have updated data
-	file := s.containerUrl.NewRootDirectoryURL().NewFileURL(name)
-	resp, err := file.Download(ctx, 0, int64(len(data)))
+	fileClient := s.containerClient.NewFileClient(name)
+	resp, err := fileClient.DownloadStream(ctx, &file.DownloadStreamOptions{
+		Range: &file.HTTPRange{Offset: 0, Count: int64(len(data))},
+	})
 	s.assert.Nil(err)
-	output, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	output, _ := io.ReadAll(resp.Body)
 	s.assert.EqualValues(testData, output)
 }
 
@@ -1562,16 +1582,17 @@ func (s *datalakeTestSuite) TestCreateLink() {
 	s.assert.Nil(err)
 
 	// Link should be in the account
-	link := s.containerUrl.NewRootDirectoryURL().NewFileURL(name)
-	props, err := link.GetProperties(ctx)
+	link := s.containerClient.NewFileClient(name)
+	props, err := link.GetProperties(ctx, nil)
 	s.assert.Nil(err)
 	s.assert.NotNil(props)
-	metadata := newMetadata(props.XMsProperties())
-	s.assert.NotEmpty(metadata)
-	s.assert.True(checkMetadata(metadata, symlinkKey, "true"))
-	resp, err := link.Download(ctx, 0, props.ContentLength())
+	s.assert.NotEmpty(props.Metadata)
+	s.assert.True(checkMetadata(props.Metadata, symlinkKey, "true"))
+	resp, err := link.DownloadStream(ctx, &file.DownloadStreamOptions{
+		Range: &file.HTTPRange{Offset: 0, Count: *props.ContentLength},
+	})
 	s.assert.Nil(err)
-	data, _ := io.ReadAll(resp.Body(azbfs.RetryReaderOptions{}))
+	data, _ := io.ReadAll(resp.Body)
 	s.assert.EqualValues(target, data)
 }
 
@@ -1701,10 +1722,11 @@ func (s *datalakeTestSuite) TestChmod() {
 	s.assert.Nil(err)
 
 	// File's ACL info should have changed
-	file := s.containerUrl.NewRootDirectoryURL().NewFileURL(name)
-	acl, err := file.GetAccessControl(ctx)
+	file := s.containerClient.NewFileClient(name)
+	acl, err := file.GetAccessControl(ctx, nil)
 	s.assert.Nil(err)
-	s.assert.EqualValues("user::rw-,group::rw-,other::rw-", acl.ACL)
+	s.assert.NotNil(acl.ACL)
+	s.assert.EqualValues("user::rw-,group::rw-,other::rw-", *acl.ACL)
 }
 
 func (s *datalakeTestSuite) TestChmodError() {
