+++ component/azstorage/utils.go
@@ -34,9 +34,7 @@
 package azstorage
 
 import (
-	"context"
 	"crypto/md5"
-	"encoding/base64"
 	"encoding/json"
 	"errors"
 	"fmt"
@@ -53,7 +51,6 @@ import (
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/cloud"
 	azlog "github.com/Azure/azure-sdk-for-go/sdk/azcore/log"
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore/policy"
-	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/blob"
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/bloberror"
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azblob/service"
@@ -63,10 +60,6 @@ import (
 	"github.com/Azure/azure-storage-fuse/v2/internal"
 
 	"github.com/Azure/azure-sdk-for-go/sdk/storage/azdatalake/datalakeerror"
-	"github.com/Azure/azure-storage-azcopy/v10/azbfs"
-	"github.com/Azure/azure-storage-azcopy/v10/ste"
-
-	"github.com/Azure/azure-pipeline-go/pipeline"
 )
 
 //    ----------- Helper to create pipeline options ---------------
@@ -88,45 +81,6 @@ const (
 	DisableCompression     bool          = false
 	MaxResponseHeaderBytes int64         = 0
 )
-
-// TODO:: track2 : remove
-// getAzBfsPipelineOptions : Create pipeline options based on the config
-func getAzBfsPipelineOptions(conf AzStorageConfig) (azbfs.PipelineOptions, ste.XferRetryOptions) {
-	retryOptions := ste.XferRetryOptions{
-		Policy:        ste.RetryPolicyExponential,                      // Use exponential backoff as opposed to linear
-		MaxTries:      conf.maxRetries,                                 // Try at most 3 times to perform the operation (set to 1 to disable retries)
-		TryTimeout:    time.Second * time.Duration(conf.maxTimeout),    // Maximum time allowed for any single try
-		RetryDelay:    time.Second * time.Duration(conf.backoffTime),   // Backoff amount for each retry (exponential or linear)
-		MaxRetryDelay: time.Second * time.Duration(conf.maxRetryDelay), // Max delay between retries
-	}
-
-	telemetryValue := conf.telemetry
-	if telemetryValue != "" {
-		telemetryValue += " "
-	}
-
-	telemetryValue += UserAgent() + " (" + common.GetCurrentDistro() + ")"
-	telemetryOptions := azbfs.TelemetryOptions{
-		Value: telemetryValue,
-	}
-
-	sysLogDisabled := log.GetType() == "silent" // If logging is enabled, allow the SDK to log retries to syslog.
-	requestLogOptions := azbfs.RequestLogOptions{
-		// TODO: We can potentially consider making LogWarningIfTryOverThreshold a user settable option. For now lets use the default
-		SyslogDisabled: sysLogDisabled,
-	}
-	logOptions := getLogOptions(conf.sdkTrace)
-	// Create custom HTTPClient to pass to the factory in order to set our proxy
-	var pipelineHTTPClient = newBlobfuse2HttpClient(&conf)
-	return azbfs.PipelineOptions{
-			Log:        logOptions,
-			RequestLog: requestLogOptions,
-			Telemetry:  telemetryOptions,
-			HTTPSender: newBlobfuse2HTTPClientFactory(pipelineHTTPClient),
-		},
-		// Set RetryOptions to control how HTTP request are retried when retryable failures occur
-		retryOptions
-}
 
 // getAzStorageClientOptions : Create client options based on the config
 func getAzStorageClientOptions(conf *AzStorageConfig) azcore.ClientOptions {
@@ -241,83 +195,6 @@ func newBlobfuse2HttpClient(conf *AzStorageConfig) *http.Client {
 	}
 }
 
-// TODO:: track2 : remove
-// newBlobfuse2HTTPClientFactory creates a custom HTTPClientPolicyFactory object that sends HTTP requests to the http client.
-func newBlobfuse2HTTPClientFactory(pipelineHTTPClient *http.Client) pipeline.Factory {
-	return pipeline.FactoryFunc(func(next pipeline.Policy, po *pipeline.PolicyOptions) pipeline.PolicyFunc {
-		return func(ctx context.Context, request pipeline.Request) (pipeline.Response, error) {
-			r, err := pipelineHTTPClient.Do(request.WithContext(ctx))
-			if err != nil {
-				log.Err("BlockBlob::newBlobfuse2HTTPClientFactory : HTTP request failed [%s]", err.Error())
-				err = pipeline.NewError(err, "HTTP request failed")
-			}
-			return pipeline.NewHTTPResponse(r), err
-		}
-	})
-}
-
-// TODO:: track2 : remove
-func getLogOptions(sdkLogging bool) pipeline.LogOptions {
-	return pipeline.LogOptions{
-		Log: func(logLevel pipeline.LogLevel, message string) {
-			if !sdkLogging {
-				return
-			}
-
-			// message here is a log generated by SDK and it contains URLs as well
-			// These URLs have '%' as part of their data like / replaced with %2F
-			// If we pass down message as first argument to our logging api, it assumes it to be
-			// a format specifier and treat each % in URL to be a type specifier this will
-			// result into log strings saying we have given %d but no integer as argument.
-			// Only way to bypass this is to pass message as a second argument to logging method
-			// so that logging api does not treat it as format string.
-			switch logLevel {
-			case pipeline.LogFatal:
-				log.Crit("SDK : %s", message)
-			case pipeline.LogPanic:
-				log.Crit("SDK : %s", message)
-			case pipeline.LogError:
-				log.Err("SDK : %s", message)
-			case pipeline.LogWarning:
-				log.Warn("SDK : %s", message)
-			case pipeline.LogInfo:
-				log.Info("SDK : %s", message)
-			case pipeline.LogDebug:
-				log.Debug("SDK : %s", message)
-			case pipeline.LogNone:
-			default:
-			}
-		},
-		ShouldLog: func(level pipeline.LogLevel) bool {
-			if !sdkLogging {
-				return false
-			}
-			currentLogLevel := func(commonLog common.LogLevel) pipeline.LogLevel {
-				switch commonLog {
-				case common.ELogLevel.INVALID():
-					return pipeline.LogNone
-				case common.ELogLevel.LOG_OFF():
-					return pipeline.LogNone
-				case common.ELogLevel.LOG_CRIT():
-					return pipeline.LogPanic // Panic logs both Panic and Fatal
-				case common.ELogLevel.LOG_ERR():
-					return pipeline.LogError
-				case common.ELogLevel.LOG_WARNING():
-					return pipeline.LogWarning
-				case common.ELogLevel.LOG_INFO():
-					return pipeline.LogInfo
-				case common.ELogLevel.LOG_TRACE():
-					return pipeline.LogDebug // No Trace in pipeline.LogLevel
-				case common.ELogLevel.LOG_DEBUG():
-					return pipeline.LogDebug
-				}
-				return pipeline.LogNone
-			}(log.GetLogLevel())
-			return level <= currentLogLevel
-		},
-	}
-}
-
 // getCloudConfiguration : returns cloud configuration type on the basis of endpoint
 func getCloudConfiguration(endpoint string) cloud.Configuration {
 	if strings.Contains(endpoint, "core.chinacloudapi.cn") {
@@ -414,32 +291,6 @@ func storeDatalakeErrToErr(err error) uint16 {
 
 //	----------- Metadata handling  ---------------
 //
-// Converts datalake properties to a metadata map
-func newMetadata(properties string) map[string]*string {
-	metadata := make(map[string]*string)
-	if properties != "" {
-		// Create a map of the properties (metadata)
-		pairs := strings.Split(properties, ",")
-
-		for _, p := range pairs {
-			components := strings.SplitN(p, "=", 2)
-			key := components[0]
-			value, err := base64.StdEncoding.DecodeString(components[1])
-			if err == nil {
-				metadata[key] = to.Ptr(string(value))
-			}
-		}
-	}
-	return metadata
-}
-
-// parseProperties : Parse the properties of a given datalake path and populate its attributes
-func parseProperties(attr *internal.ObjAttr, properties string) {
-	metadata := newMetadata(properties)
-	// Parse the metadata
-	parseMetadata(attr, metadata)
-}
-
 // parseMetadata : Parse the metadata of a given path and populate its attributes
 func parseMetadata(attr *internal.ObjAttr, metadata map[string]*string) {
 	// Save the metadata in attributes so that later if someone wants to add anything it can work
@@ -543,8 +394,6 @@ func populateContentType(newSet string) error { //nolint
 	return nil
 }
 
-// TODO:: track2 : review : added "cold" and "premium" tiers
-//
 //	----------- Blob access tier type conversion  ---------------
 //
 // AccessTierMap : Store config to access tier mapping
